requirements.txt
numpy
pandas
scipy.stats
streamlit
matplotlib.pyplot
io